+++ compiler/scenario-service/server/src/main/scala/com/digitalasset/daml/lf/scenario/Conversions.scala
@@ -4,25 +4,25 @@
 package com.daml.lf.scenario
 
 import com.daml.lf.data.{Numeric, Ref}
-import com.daml.lf.scenario.api.v1
-import com.daml.lf.scenario.api.v1.{List => _, _}
+import com.daml.lf.ledger.EventId
+import com.daml.lf.scenario.api.{v1 => proto}
 import com.daml.lf.speedy.{SError, SValue, Speedy, PartialTransaction => SPartialTransaction}
 import com.daml.lf.transaction.{Node => N, Transaction => Tx}
-import com.daml.lf.types.Ledger
+import com.daml.lf.ledger._
 import com.daml.lf.value.{Value => V}
 
 import scala.collection.JavaConverters._
 
 final class Conversions(
     homePackageId: Ref.PackageId,
-    ledger: Ledger.Ledger,
+    ledger: ScenarioLedger,
     machine: Speedy.Machine,
 ) {
 
-  private val empty: Empty = Empty.newBuilder.build
+  private val empty: proto.Empty = proto.Empty.newBuilder.build
 
-  private val packageIdSelf: PackageIdentifier =
-    PackageIdentifier.newBuilder.setSelf(empty).build
+  private val packageIdSelf: proto.PackageIdentifier =
+    proto.PackageIdentifier.newBuilder.setSelf(empty).build
 
   // The ledger data will not contain information from the partial transaction at this point.
   // We need the mapping for converting error message so we manually add it here.
@@ -41,8 +41,8 @@ final class Conversions(
     case (idx, step) => convertScenarioStep(idx.toInt, step)
   }
 
-  def convertScenarioResult(svalue: SValue): ScenarioResult = {
-    val builder = ScenarioResult.newBuilder
+  def convertScenarioResult(svalue: SValue): proto.ScenarioResult = {
+    val builder = proto.ScenarioResult.newBuilder
       .addAllNodes(nodes.asJava)
       .addAllScenarioSteps(steps.asJava)
       .setReturnValue(convertSValue(svalue))
@@ -55,8 +55,8 @@ final class Conversions(
 
   def convertScenarioError(
       err: SError.SError,
-  ): ScenarioError = {
-    val builder = ScenarioError.newBuilder
+  ): proto.ScenarioError = {
+    val builder = proto.ScenarioError.newBuilder
       .addAllNodes(nodes.asJava)
       .addAllScenarioSteps(steps.asJava)
       .setLedgerTime(ledger.currentTime.micros)
@@ -91,7 +91,7 @@ final class Conversions(
         setCrash(reason)
 
       case SError.DamlETemplatePreconditionViolated(tid, optLoc, arg) =>
-        val uepvBuilder = ScenarioError.TemplatePreconditionViolated.newBuilder
+        val uepvBuilder = proto.ScenarioError.TemplatePreconditionViolated.newBuilder
         optLoc.map(convertLocation).foreach(uepvBuilder.setLocation)
         builder.setTemplatePrecondViolated(
           uepvBuilder
@@ -101,14 +101,14 @@ final class Conversions(
         )
       case SError.DamlELocalContractNotActive(coid, tid, consumedBy) =>
         builder.setUpdateLocalContractNotActive(
-          ScenarioError.ContractNotActive.newBuilder
+          proto.ScenarioError.ContractNotActive.newBuilder
             .setContractRef(mkContractRef(coid, tid))
-            .setConsumedBy(NodeId.newBuilder.setId(consumedBy.toString).build)
+            .setConsumedBy(proto.NodeId.newBuilder.setId(consumedBy.toString).build)
             .build,
         )
       case SError.ScenarioErrorContractNotEffective(coid, tid, effectiveAt) =>
         builder.setScenarioContractNotEffective(
-          ScenarioError.ContractNotEffective.newBuilder
+          proto.ScenarioError.ContractNotEffective.newBuilder
             .setEffectiveAt(effectiveAt.micros)
             .setContractRef(mkContractRef(coid, tid))
             .build,
@@ -116,7 +116,7 @@ final class Conversions(
 
       case SError.ScenarioErrorContractNotActive(coid, tid, consumedBy) =>
         builder.setScenarioContractNotActive(
-          ScenarioError.ContractNotActive.newBuilder
+          proto.ScenarioError.ContractNotActive.newBuilder
             .setContractRef(mkContractRef(coid, tid))
             .setConsumedBy(convertNodeId(consumedBy))
             .build,
@@ -124,7 +124,7 @@ final class Conversions(
 
       case SError.ScenarioErrorContractNotVisible(coid, tid, committer, observers) =>
         builder.setScenarioContractNotVisible(
-          ScenarioError.ContractNotVisible.newBuilder
+          proto.ScenarioError.ContractNotVisible.newBuilder
             .setContractRef(mkContractRef(coid, tid))
             .setCommitter(convertParty(committer))
             .addAllObservers(observers.map(convertParty).asJava)
@@ -133,7 +133,7 @@ final class Conversions(
 
       case SError.ScenarioErrorContractKeyNotVisible(coid, gk, committer, observers) =>
         builder.setScenarioContractKeyNotVisible(
-          ScenarioError.ContractKeyNotVisible.newBuilder
+          proto.ScenarioError.ContractKeyNotVisible.newBuilder
             .setContractRef(mkContractRef(coid, gk.templateId))
             .setKey(convertValue(gk.key))
             .setCommitter(convertParty(committer))
@@ -159,27 +159,27 @@ final class Conversions(
     builder.build
   }
 
-  def convertCommitError(commitError: Ledger.CommitError): CommitError = {
-    val builder = CommitError.newBuilder
+  def convertCommitError(commitError: ScenarioLedger.CommitError): proto.CommitError = {
+    val builder = proto.CommitError.newBuilder
     commitError match {
-      case Ledger.CommitError.UniqueKeyViolation(gk) =>
+      case ScenarioLedger.CommitError.UniqueKeyViolation(gk) =>
         builder.setUniqueKeyViolation(convertGlobalKey(gk.gk))
-      case Ledger.CommitError.FailedAuthorizations(fas) =>
+      case ScenarioLedger.CommitError.FailedAuthorizations(fas) =>
         builder.setFailedAuthorizations(convertFailedAuthorizations(fas))
     }
     builder.build
   }
 
-  def convertGlobalKey(globalKey: N.GlobalKey): GlobalKey = {
-    GlobalKey.newBuilder
+  def convertGlobalKey(globalKey: N.GlobalKey): proto.GlobalKey = {
+    proto.GlobalKey.newBuilder
       .setTemplateId(convertIdentifier(globalKey.templateId))
       .setKey(convertValue(globalKey.key))
       .build
   }
 
-  def convertSValue(svalue: SValue): Value = {
-    def unserializable(what: String): Value =
-      Value.newBuilder.setUnserializable(what).build
+  def convertSValue(svalue: SValue): proto.Value = {
+    def unserializable(what: String): proto.Value =
+      proto.Value.newBuilder.setUnserializable(what).build
     try {
       convertValue(svalue.toValue)
     } catch {
@@ -195,58 +195,58 @@ final class Conversions(
     }
   }
 
-  def convertSTraceMessage(msgAndLoc: (String, Option[Ref.Location])): TraceMessage = {
-    val builder = TraceMessage.newBuilder
+  def convertSTraceMessage(msgAndLoc: (String, Option[Ref.Location])): proto.TraceMessage = {
+    val builder = proto.TraceMessage.newBuilder
     msgAndLoc._2.map(loc => builder.setLocation(convertLocation(loc)))
     builder.setMessage(msgAndLoc._1).build
   }
 
-  def convertFailedAuthorizations(fas: Ledger.FailedAuthorizations): FailedAuthorizations = {
-    val builder = FailedAuthorizations.newBuilder
+  def convertFailedAuthorizations(fas: FailedAuthorizations): proto.FailedAuthorizations = {
+    val builder = proto.FailedAuthorizations.newBuilder
     fas.map {
       case (nodeId, fa) =>
         builder.addFailedAuthorizations {
-          val faBuilder = FailedAuthorization.newBuilder
+          val faBuilder = proto.FailedAuthorization.newBuilder
           faBuilder.setNodeId(convertTxNodeId(nodeId))
           fa match {
-            case Ledger.FACreateMissingAuthorization(
+            case FailedAuthorization.CreateMissingAuthorization(
                 templateId,
                 optLocation,
                 authParties,
                 reqParties,
                 ) =>
               val cmaBuilder =
-                FailedAuthorization.CreateMissingAuthorization.newBuilder
+                proto.FailedAuthorization.CreateMissingAuthorization.newBuilder
                   .setTemplateId(convertIdentifier(templateId))
                   .addAllAuthorizingParties(authParties.map(convertParty).asJava)
                   .addAllRequiredAuthorizers(reqParties.map(convertParty).asJava)
               optLocation.map(loc => cmaBuilder.setLocation(convertLocation(loc)))
               faBuilder.setCreateMissingAuthorization(cmaBuilder.build)
 
-            case Ledger.FAMaintainersNotSubsetOfSignatories(
+            case FailedAuthorization.MaintainersNotSubsetOfSignatories(
                 templateId,
                 optLocation,
                 signatories,
                 maintainers,
                 ) =>
               val maintNotSignBuilder =
-                FailedAuthorization.MaintainersNotSubsetOfSignatories.newBuilder
+                proto.FailedAuthorization.MaintainersNotSubsetOfSignatories.newBuilder
                   .setTemplateId(convertIdentifier(templateId))
                   .addAllSignatories(signatories.map(convertParty).asJava)
                   .addAllMaintainers(maintainers.map(convertParty).asJava)
               optLocation.map(loc => maintNotSignBuilder.setLocation(convertLocation(loc)))
               faBuilder.setMaintainersNotSubsetOfSignatories(maintNotSignBuilder.build)
 
-            case fma: Ledger.FAFetchMissingAuthorization =>
+            case fma: FailedAuthorization.FetchMissingAuthorization =>
               val fmaBuilder =
-                FailedAuthorization.FetchMissingAuthorization.newBuilder
+                proto.FailedAuthorization.FetchMissingAuthorization.newBuilder
                   .setTemplateId(convertIdentifier(fma.templateId))
                   .addAllAuthorizingParties(fma.authorizingParties.map(convertParty).asJava)
                   .addAllStakeholders(fma.stakeholders.map(convertParty).asJava)
               fma.optLocation.map(loc => fmaBuilder.setLocation(convertLocation(loc)))
               faBuilder.setFetchMissingAuthorization(fmaBuilder.build)
 
-            case Ledger.FAExerciseMissingAuthorization(
+            case FailedAuthorization.ExerciseMissingAuthorization(
                 templateId,
                 choiceId,
                 optLocation,
@@ -254,44 +254,48 @@ final class Conversions(
                 reqParties,
                 ) =>
               val emaBuilder =
-                FailedAuthorization.ExerciseMissingAuthorization.newBuilder
+                proto.FailedAuthorization.ExerciseMissingAuthorization.newBuilder
                   .setTemplateId(convertIdentifier(templateId))
                   .setChoiceId(choiceId)
                   .addAllAuthorizingParties(authParties.map(convertParty).asJava)
                   .addAllRequiredAuthorizers(reqParties.map(convertParty).asJava)
               optLocation.map(loc => emaBuilder.setLocation(convertLocation(loc)))
               faBuilder.setExerciseMissingAuthorization(emaBuilder.build)
-            case Ledger.FAActorMismatch(templateId, choiceId, optLocation, givenActors) =>
+            case FailedAuthorization.ActorMismatch(
+                templateId,
+                choiceId,
+                optLocation,
+                givenActors) =>
               val amBuilder =
-                FailedAuthorization.ActorMismatch.newBuilder
+                proto.FailedAuthorization.ActorMismatch.newBuilder
                   .setTemplateId(convertIdentifier(templateId))
                   .setChoiceId(choiceId)
                   .addAllGivenActors(givenActors.map(convertParty).asJava)
               optLocation.map(loc => amBuilder.setLocation(convertLocation(loc)))
               faBuilder.setActorMismatch(amBuilder.build)
-            case Ledger.FANoSignatories(templateId, optLocation) =>
+            case FailedAuthorization.NoSignatories(templateId, optLocation) =>
               val nsBuilder =
-                FailedAuthorization.NoSignatories.newBuilder
+                proto.FailedAuthorization.NoSignatories.newBuilder
                   .setTemplateId(convertIdentifier(templateId))
               optLocation.map(loc => nsBuilder.setLocation(convertLocation(loc)))
               faBuilder.setNoSignatories(nsBuilder.build)
 
-            case Ledger.FANoControllers(templateId, choiceId, optLocation) =>
+            case FailedAuthorization.NoControllers(templateId, choiceId, optLocation) =>
               val ncBuilder =
-                FailedAuthorization.NoControllers.newBuilder
+                proto.FailedAuthorization.NoControllers.newBuilder
                   .setTemplateId(convertIdentifier(templateId))
                   .setChoiceId(choiceId)
               optLocation.map(loc => ncBuilder.setLocation(convertLocation(loc)))
               faBuilder.setNoControllers(ncBuilder.build)
 
-            case Ledger.FALookupByKeyMissingAuthorization(
+            case FailedAuthorization.LookupByKeyMissingAuthorization(
                 templateId,
                 optLocation,
                 maintainers,
                 authorizers,
                 ) =>
               val lbkmaBuilder =
-                FailedAuthorization.LookupByKeyMissingAuthorization.newBuilder
+                proto.FailedAuthorization.LookupByKeyMissingAuthorization.newBuilder
                   .setTemplateId(convertIdentifier(templateId))
                   .addAllMaintainers(maintainers.map(convertParty).asJava)
                   .addAllAuthorizingParties(authorizers.map(convertParty).asJava)
@@ -304,8 +308,8 @@ final class Conversions(
     builder.build
   }
 
-  def mkContractRef(coid: V.ContractId, templateId: Ref.Identifier): ContractRef =
-    ContractRef.newBuilder
+  def mkContractRef(coid: V.ContractId, templateId: Ref.Identifier): proto.ContractRef =
+    proto.ContractRef.newBuilder
       .setRelative(false)
       .setContractId(coidToEventId(coid).toLedgerString)
       .setTemplateId(convertIdentifier(templateId))
@@ -313,13 +317,13 @@ final class Conversions(
 
   def convertScenarioStep(
       stepId: Int,
-      step: Ledger.ScenarioStep,
-  ): ScenarioStep = {
-    val builder = ScenarioStep.newBuilder
+      step: ScenarioLedger.ScenarioStep,
+  ): proto.ScenarioStep = {
+    val builder = proto.ScenarioStep.newBuilder
     builder.setStepId(stepId)
     step match {
-      case Ledger.Commit(txId, rtx, optLocation) =>
-        val commitBuilder = ScenarioStep.Commit.newBuilder
+      case ScenarioLedger.Commit(txId, rtx, optLocation) =>
+        val commitBuilder = proto.ScenarioStep.Commit.newBuilder
         optLocation.map { loc =>
           commitBuilder.setLocation(convertLocation(loc))
         }
@@ -329,10 +333,10 @@ final class Conversions(
             .setTx(convertTransaction(rtx))
             .build,
         )
-      case Ledger.PassTime(dt) =>
+      case ScenarioLedger.PassTime(dt) =>
         builder.setPassTime(dt)
-      case Ledger.AssertMustFail(actor, optLocation, time, txId) =>
-        val assertBuilder = ScenarioStep.AssertMustFail.newBuilder
+      case ScenarioLedger.AssertMustFail(actor, optLocation, time, txId) =>
+        val assertBuilder = proto.ScenarioStep.AssertMustFail.newBuilder
         optLocation.map { loc =>
           assertBuilder.setLocation(convertLocation(loc))
         }
@@ -349,12 +353,12 @@ final class Conversions(
   }
 
   def convertTransaction(
-      rtx: Ledger.RichTransaction,
-  ): Transaction = {
+      rtx: ScenarioLedger.RichTransaction,
+  ): proto.Transaction = {
     val convertedGlobalImplicitDisclosure = rtx.globalImplicitDisclosure.map {
       case (coid, parties) => coidToEventId(coid) -> parties
     }
-    Transaction.newBuilder
+    proto.Transaction.newBuilder
       .setCommitter(convertParty(rtx.committer))
       .setEffectiveAt(rtx.effectiveAt.micros)
       .addAllRoots(rtx.roots.map(convertNodeId).toSeq.asJava)
@@ -363,7 +367,7 @@ final class Conversions(
       // therefore we need to explicitly add the contracts that are divulged directly (via ContractId rather than ScenarioNodeId)
       .addAllDisclosures((rtx.disclosures ++ convertedGlobalImplicitDisclosure).toSeq.map {
         case (nodeId, parties) =>
-          NodeAndParties.newBuilder
+          proto.NodeAndParties.newBuilder
             .setNodeId(convertNodeId(nodeId))
             .addAllParties(parties.map(convertParty).asJava)
             .build
@@ -372,8 +376,8 @@ final class Conversions(
       .build
   }
 
-  def convertPartialTransaction(ptx: SPartialTransaction): PartialTransaction = {
-    val builder = PartialTransaction.newBuilder
+  def convertPartialTransaction(ptx: SPartialTransaction): proto.PartialTransaction = {
+    val builder = proto.PartialTransaction.newBuilder
       .addAllNodes(ptx.nodes.map(convertNode(convertValue(_), _)).asJava)
       .addAllRoots(
         ptx.context.children.toImmArray.toSeq.sortBy(_.index).map(convertTxNodeId).asJava,
@@ -382,7 +386,7 @@ final class Conversions(
     ptx.context.exeContext match {
       case None =>
       case Some(ctx) =>
-        val ecBuilder = ExerciseContext.newBuilder
+        val ecBuilder = proto.ExerciseContext.newBuilder
           .setTargetId(mkContractRef(ctx.targetId, ctx.templateId))
           .setChoiceId(ctx.choiceId)
           .setChosenValue(convertValue(ctx.chosenValue))
@@ -392,21 +396,21 @@ final class Conversions(
     builder.build
   }
 
-  def convertNodeId(nodeId: Ledger.EventId): NodeId =
-    NodeId.newBuilder.setId(nodeId.toLedgerString).build
+  def convertNodeId(nodeId: EventId): proto.NodeId =
+    proto.NodeId.newBuilder.setId(nodeId.toLedgerString).build
 
-  def convertTxNodeId(nodeId: Tx.NodeId): NodeId =
-    NodeId.newBuilder.setId(nodeId.index.toString).build
+  def convertTxNodeId(nodeId: Tx.NodeId): proto.NodeId =
+    proto.NodeId.newBuilder.setId(nodeId.index.toString).build
 
-  def convertNode(nodeId: Ledger.EventId, nodeInfo: Ledger.LedgerNodeInfo): Node = {
-    val builder = Node.newBuilder
+  def convertNode(nodeId: EventId, nodeInfo: ScenarioLedger.LedgerNodeInfo): proto.Node = {
+    val builder = proto.Node.newBuilder
     builder
       .setNodeId(convertNodeId(nodeId))
       .setEffectiveAt(nodeInfo.effectiveAt.micros)
       .addAllReferencedBy(nodeInfo.referencedBy.map(convertNodeId).asJava)
       .addAllObservingSince(nodeInfo.observingSince.toList.map {
         case (party, txId) =>
-          PartyAndTransactionId.newBuilder
+          proto.PartyAndTransactionId.newBuilder
             .setParty(convertParty(party))
             .setTxId(txId.index)
             .build
@@ -420,9 +424,9 @@ final class Conversions(
     nodeInfo.node match {
       case create: N.NodeCreate[V.ContractId, Tx.Value[V.ContractId]] =>
         val createBuilder =
-          Node.Create.newBuilder
+          proto.Node.Create.newBuilder
             .setContractInstance(
-              ContractInstance.newBuilder
+              proto.ContractInstance.newBuilder
                 .setTemplateId(convertIdentifier(create.coinst.template))
                 .setValue(convertValue(create.coinst.arg.value))
                 .build,
@@ -434,7 +438,7 @@ final class Conversions(
         builder.setCreate(createBuilder.build)
       case fetch: N.NodeFetch.WithTxValue[V.ContractId] =>
         builder.setFetch(
-          Node.Fetch.newBuilder
+          proto.Node.Fetch.newBuilder
             .setContractId(coidToEventId(fetch.coid).toLedgerString)
             .setTemplateId(convertIdentifier(fetch.templateId))
             .addAllSignatories(fetch.signatories.map(convertParty).asJava)
@@ -442,14 +446,14 @@ final class Conversions(
             .build,
         )
       case ex: N.NodeExercises[
-            Ledger.EventId,
+            EventId,
             V.ContractId,
             Tx.Value[
               V.ContractId,
             ]] =>
         ex.optLocation.map(loc => builder.setLocation(convertLocation(loc)))
         builder.setExercise(
-          Node.Exercise.newBuilder
+          proto.Node.Exercise.newBuilder
             .setTargetContractId(coidToEventId(ex.targetCoid).toLedgerString)
             .setTemplateId(convertIdentifier(ex.templateId))
             .setChoiceId(ex.choiceId)
@@ -460,7 +464,7 @@ final class Conversions(
             .addAllStakeholders(ex.stakeholders.map(convertParty).asJava)
             .addAllChildren(
               ex.children
-                .map(nid => NodeId.newBuilder.setId(nid.toLedgerString).build)
+                .map(nid => proto.NodeId.newBuilder.setId(nid.toLedgerString).build)
                 .toSeq
                 .asJava,
             )
@@ -469,7 +473,7 @@ final class Conversions(
 
       case lbk: N.NodeLookupByKey[V.ContractId, Tx.Value[V.ContractId]] =>
         lbk.optLocation.foreach(loc => builder.setLocation(convertLocation(loc)))
-        val lbkBuilder = Node.LookupByKey.newBuilder
+        val lbkBuilder = proto.Node.LookupByKey.newBuilder
           .setTemplateId(convertIdentifier(lbk.templateId))
           .setKeyWithMaintainers(convertKeyWithMaintainers(convertVersionedValue, lbk.key))
         lbk.result.foreach(cid => lbkBuilder.setContractId(coidToEventId(cid).toLedgerString))
@@ -480,10 +484,10 @@ final class Conversions(
   }
 
   def convertKeyWithMaintainers[Val](
-      convertValue: Val => Value,
+      convertValue: Val => proto.Value,
       key: N.KeyWithMaintainers[Val],
-  ): KeyWithMaintainers = {
-    KeyWithMaintainers
+  ): proto.KeyWithMaintainers = {
+    proto.KeyWithMaintainers
       .newBuilder()
       .setKey(convertValue(key.key))
       .addAllMaintainers(key.maintainers.map(convertParty).asJava)
@@ -491,20 +495,20 @@ final class Conversions(
   }
 
   def convertNode[Val](
-      convertValue: Val => Value,
+      convertValue: Val => proto.Value,
       nodeWithId: (Tx.NodeId, N.GenNode[V.NodeId, V.ContractId, Val]),
-  ): Node = {
+  ): proto.Node = {
     val (nodeId, node) = nodeWithId
-    val builder = Node.newBuilder
+    val builder = proto.Node.newBuilder
     builder
-      .setNodeId(NodeId.newBuilder.setId(nodeId.index.toString).build)
+      .setNodeId(proto.NodeId.newBuilder.setId(nodeId.index.toString).build)
     // FIXME(JM): consumedBy, parent, ...
     node match {
       case create: N.NodeCreate[V.ContractId, Val] =>
         val createBuilder =
-          Node.Create.newBuilder
+          proto.Node.Create.newBuilder
             .setContractInstance(
-              ContractInstance.newBuilder
+              proto.ContractInstance.newBuilder
                 .setTemplateId(convertIdentifier(create.coinst.template))
                 .setValue(convertValue(create.coinst.arg))
                 .build,
@@ -517,7 +521,7 @@ final class Conversions(
         builder.setCreate(createBuilder.build)
       case fetch: N.NodeFetch[V.ContractId, Val] =>
         builder.setFetch(
-          Node.Fetch.newBuilder
+          proto.Node.Fetch.newBuilder
             .setContractId(coidToEventId(fetch.coid).toLedgerString)
             .setTemplateId(convertIdentifier(fetch.templateId))
             .addAllSignatories(fetch.signatories.map(convertParty).asJava)
@@ -527,7 +531,7 @@ final class Conversions(
       case ex: N.NodeExercises[V.NodeId, V.ContractId, Val] =>
         ex.optLocation.map(loc => builder.setLocation(convertLocation(loc)))
         builder.setExercise(
-          Node.Exercise.newBuilder
+          proto.Node.Exercise.newBuilder
             .setTargetContractId(coidToEventId(ex.targetCoid).toLedgerString)
             .setTemplateId(convertIdentifier(ex.templateId))
             .setChoiceId(ex.choiceId)
@@ -538,7 +542,7 @@ final class Conversions(
             .addAllStakeholders(ex.stakeholders.map(convertParty).asJava)
             .addAllChildren(
               ex.children
-                .map(nid => NodeId.newBuilder.setId(nid.index.toString).build)
+                .map(nid => proto.NodeId.newBuilder.setId(nid.index.toString).build)
                 .toSeq
                 .asJava,
             )
@@ -548,7 +552,7 @@ final class Conversions(
       case lookup: N.NodeLookupByKey[V.ContractId, Val] =>
         lookup.optLocation.map(loc => builder.setLocation(convertLocation(loc)))
         builder.setLookupByKey({
-          val builder = Node.LookupByKey.newBuilder
+          val builder = proto.Node.LookupByKey.newBuilder
             .setKeyWithMaintainers(convertKeyWithMaintainers(convertValue, lookup.key))
           lookup.result.foreach(cid => builder.setContractId(coidToEventId(cid).toLedgerString))
           builder.build
@@ -557,23 +561,23 @@ final class Conversions(
     builder.build
   }
 
-  def convertPackageId(pkg: Ref.PackageId): PackageIdentifier =
+  def convertPackageId(pkg: Ref.PackageId): proto.PackageIdentifier =
     if (pkg == homePackageId)
       // Reconstitute the self package reference.
       packageIdSelf
     else
-      PackageIdentifier.newBuilder.setPackageId(pkg).build
+      proto.PackageIdentifier.newBuilder.setPackageId(pkg).build
 
-  def convertIdentifier(identifier: Ref.Identifier): Identifier =
-    Identifier.newBuilder
+  def convertIdentifier(identifier: Ref.Identifier): proto.Identifier =
+    proto.Identifier.newBuilder
       .setPackage(convertPackageId(identifier.packageId))
       .setName(identifier.qualifiedName.toString)
       .build
 
-  def convertLocation(loc: Ref.Location): Location = {
+  def convertLocation(loc: Ref.Location): proto.Location = {
     val (sline, scol) = loc.start
     val (eline, ecol) = loc.end
-    Location.newBuilder
+    proto.Location.newBuilder
       .setPackage(convertPackageId(loc.packageId))
       .setModule(loc.module.toString)
       .setDefinition(loc.definition)
@@ -585,21 +589,21 @@ final class Conversions(
 
   }
 
-  private def convertVersionedValue(value: V.VersionedValue[V.ContractId]): Value =
+  private def convertVersionedValue(value: V.VersionedValue[V.ContractId]): proto.Value =
     convertValue(value.value)
 
-  def convertValue(value: V[V.ContractId]): Value = {
-    val builder = Value.newBuilder
+  def convertValue(value: V[V.ContractId]): proto.Value = {
+    val builder = proto.Value.newBuilder
     value match {
       case V.ValueRecord(tycon, fields) =>
-        val rbuilder = Record.newBuilder
+        val rbuilder = proto.Record.newBuilder
         tycon.map(x => rbuilder.setRecordId(convertIdentifier(x)))
         builder.setRecord(
           rbuilder
             .addAllFields(
               fields.toSeq.map {
                 case (optName, fieldValue) =>
-                  val builder = Field.newBuilder
+                  val builder = proto.Field.newBuilder
                   optName.foreach(builder.setLabel)
                   builder
                     .setValue(convertValue(fieldValue))
@@ -610,10 +614,10 @@ final class Conversions(
         )
       case V.ValueStruct(fields) =>
         builder.setTuple(
-          Tuple.newBuilder
+          proto.Tuple.newBuilder
             .addAllFields(
               fields.toSeq.map { field =>
-                Field.newBuilder
+                proto.Field.newBuilder
                   .setLabel(field._1)
                   .setValue(convertValue(field._2))
                   .build
@@ -622,7 +626,7 @@ final class Conversions(
             .build,
         )
       case V.ValueVariant(tycon, variant, value) =>
-        val vbuilder = Variant.newBuilder
+        val vbuilder = proto.Variant.newBuilder
         tycon.foreach(x => vbuilder.setVariantId(convertIdentifier(x)))
         builder.setVariant(
           vbuilder
@@ -631,14 +635,14 @@ final class Conversions(
             .build,
         )
       case V.ValueEnum(tycon, constructor) =>
-        val eBuilder = Enum.newBuilder.setConstructor(constructor)
+        val eBuilder = proto.Enum.newBuilder.setConstructor(constructor)
         tycon.foreach(x => eBuilder.setEnumId(convertIdentifier(x)))
         builder.setEnum(eBuilder.build)
       case V.ValueContractId(coid) =>
         builder.setContractId(coidToEventId(coid).toLedgerString)
       case V.ValueList(values) =>
         builder.setList(
-          v1.List.newBuilder
+          proto.List.newBuilder
             .addAllElements(
               values
                 .map(convertValue)
@@ -657,26 +661,26 @@ final class Conversions(
       case V.ValueBool(b) => builder.setBool(b)
       case V.ValueUnit => builder.setUnit(empty)
       case V.ValueOptional(mbV) =>
-        val optionalBuilder = Optional.newBuilder
+        val optionalBuilder = proto.Optional.newBuilder
         mbV match {
           case None => ()
           case Some(v) => optionalBuilder.setValue(convertValue(v))
         }
         builder.setOptional(optionalBuilder)
       case V.ValueTextMap(map) =>
-        val mapBuilder = v1.Map.newBuilder
+        val mapBuilder = proto.Map.newBuilder
         map.toImmArray.foreach {
           case (k, v) =>
-            mapBuilder.addEntries(v1.Map.Entry.newBuilder().setKey(k).setValue(convertValue(v)))
+            mapBuilder.addEntries(proto.Map.Entry.newBuilder().setKey(k).setValue(convertValue(v)))
             ()
         }
         builder.setMap(mapBuilder)
       case V.ValueGenMap(entries) =>
-        val mapBuilder = v1.GenMap.newBuilder
+        val mapBuilder = proto.GenMap.newBuilder
         entries.foreach {
           case (k, v) =>
             mapBuilder.addEntries(
-              v1.GenMap.Entry.newBuilder().setKey(convertValue(k)).setValue(convertValue(v)),
+              proto.GenMap.Entry.newBuilder().setKey(convertValue(k)).setValue(convertValue(v)),
             )
             ()
         }
@@ -685,7 +689,7 @@ final class Conversions(
     builder.build
   }
 
-  def convertParty(p: Ref.Party): Party =
-    Party.newBuilder.setParty(p).build
+  def convertParty(p: Ref.Party): proto.Party =
+    proto.Party.newBuilder.setParty(p).build
 
 }
