+++ daml-lf/scenario-interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/ScenarioRunner.scala
@@ -3,8 +3,7 @@
 
 package com.daml.lf.speedy
 
-import com.daml.lf.types.Ledger
-import com.daml.lf.types.Ledger._
+import com.daml.lf.scenario.ScenarioLedger
 import com.daml.lf.data.Ref._
 import com.daml.lf.data.Time
 import com.daml.lf.transaction.Transaction._
@@ -27,11 +26,11 @@ private case class SRunnerException(err: SError) extends RuntimeException(err.to
 final case class ScenarioRunner(
     machine: Speedy.Machine,
     partyNameMangler: (String => String) = identity) {
-  var ledger: Ledger = Ledger.initialLedger(Time.Timestamp.Epoch)
+  var ledger: ScenarioLedger = ScenarioLedger.initialLedger(Time.Timestamp.Epoch)
 
   import scala.util.{Try, Success, Failure}
 
-  def run(): Either[(SError, Ledger), (Double, Int, Ledger, SValue)] =
+  def run(): Either[(SError, ScenarioLedger), (Double, Int, ScenarioLedger, SValue)] =
     Try(runUnsafe) match {
       case Failure(SRunnerException(err)) =>
         Left((err, ledger))
@@ -41,7 +40,7 @@ final case class ScenarioRunner(
         Right(res)
     }
 
-  private def runUnsafe(): (Double, Int, Ledger, SValue) = {
+  private def runUnsafe(): (Double, Int, ScenarioLedger, SValue) = {
     // NOTE(JM): Written with an imperative loop and exceptions for speed
     // and so that we don't need to worry about stack usage.
     val startTime = System.nanoTime()
@@ -112,7 +111,7 @@ final case class ScenarioRunner(
     val committer =
       if (committers.size == 1) committers.head else crashTooManyCommitters(committers)
 
-    if (Ledger
+    if (ScenarioLedger
         .commitTransaction(
           committer = committer,
           effectiveAt = ledger.currentTime,
@@ -133,7 +132,7 @@ final case class ScenarioRunner(
     val committer =
       if (committers.size == 1) committers.head else crashTooManyCommitters(committers)
 
-    Ledger.commitTransaction(
+    ScenarioLedger.commitTransaction(
       committer = committer,
       effectiveAt = ledger.currentTime,
       optLocation = machine.commitLocation,
@@ -167,22 +166,25 @@ final case class ScenarioRunner(
       if (!cbMissing(()))
         throw SRunnerException(err)
 
-    ledger.lookupGlobalContract(view = ParticipantView(committer), effectiveAt = effectiveAt, acoid) match {
-      case LookupOk(_, coinst, _) =>
+    ledger.lookupGlobalContract(
+      view = ScenarioLedger.ParticipantView(committer),
+      effectiveAt = effectiveAt,
+      acoid) match {
+      case ScenarioLedger.LookupOk(_, coinst, _) =>
         cbPresent(coinst)
 
-      case LookupContractNotFound(coid) =>
+      case ScenarioLedger.LookupContractNotFound(coid) =>
         // This should never happen, hence we don't have a specific
         // error for this.
         missingWith(SErrorCrash(s"contract $coid not found"))
 
-      case LookupContractNotEffective(coid, tid, effectiveAt) =>
+      case ScenarioLedger.LookupContractNotEffective(coid, tid, effectiveAt) =>
         missingWith(ScenarioErrorContractNotEffective(coid, tid, effectiveAt))
 
-      case LookupContractNotActive(coid, tid, consumedBy) =>
+      case ScenarioLedger.LookupContractNotActive(coid, tid, consumedBy) =>
         missingWith(ScenarioErrorContractNotActive(coid, tid, consumedBy))
 
-      case LookupContractNotVisible(coid, tid, observers) =>
+      case ScenarioLedger.LookupContractNotVisible(coid, tid, observers) =>
         missingWith(ScenarioErrorContractNotVisible(coid, tid, committer, observers))
     }
   }
@@ -205,22 +207,22 @@ final case class ScenarioRunner(
         missingWith(SErrorCrash(s"Key $gk not found"))
       case Some(acoid) =>
         ledger.lookupGlobalContract(
-          view = ParticipantView(committer),
+          view = ScenarioLedger.ParticipantView(committer),
           effectiveAt = effectiveAt,
           acoid) match {
-          case LookupOk(_, _, stakeholders) =>
+          case ScenarioLedger.LookupOk(_, _, stakeholders) =>
             if (stakeholders.contains(committer))
               cb(SKeyLookupResult.Found(acoid))
             else if (!cb(SKeyLookupResult.NotVisible))
               missingWith(ScenarioErrorContractKeyNotVisible(acoid, gk, committer, stakeholders))
             ()
-          case LookupContractNotFound(coid) =>
+          case ScenarioLedger.LookupContractNotFound(coid) =>
             missingWith(SErrorCrash(s"contract $coid not found, but we found its key!"))
-          case LookupContractNotEffective(_, _, _) =>
+          case ScenarioLedger.LookupContractNotEffective(_, _, _) =>
             missingWith(SErrorCrash(s"contract $acoid not effective, but we found its key!"))
-          case LookupContractNotActive(_, _, _) =>
+          case ScenarioLedger.LookupContractNotActive(_, _, _) =>
             missingWith(SErrorCrash(s"contract $acoid not active, but we found its key!"))
-          case LookupContractNotVisible(coid, tid, observers) =>
+          case ScenarioLedger.LookupContractNotVisible(coid, tid, observers) =>
             if (!cb(SKeyLookupResult.NotVisible))
               missingWith(ScenarioErrorContractKeyNotVisible(coid, gk, committer, observers))
         }
