+++ daml-lf/transaction/src/main/scala/com/digitalasset/daml/lf/ledger/package.scala
@@ -1,0 +1,7 @@
+package com.daml.lf
+
+package object ledger {
+
+  type FailedAuthorizations = Map[transaction.Transaction.NodeId, FailedAuthorization]
+
+}
